Error_Grande_Solucionado

+++ src/main/java/com/comuctiva/comuctiva/models/Comp_ProducId.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
     public class Comp_ProducId {
-        private Long compraId;
+        private Integer compraId;
         private Long producId;
 }
