+++ src/main/java/com/comuctiva/comuctiva/models/Produc_CarriId.java
@@ -13,6 +13,6 @@ import lombok.NoArgsConstructor;
 @Data
 public class Produc_CarriId implements Serializable{
     private Long prodId;
-    private Long carritoId;
+    private Integer carritoId;
 }
